requirements.txt
Django>=5.0.7,<=5.1
djangorestframework>=3.14.0,<3.15
python-decouple==3.8
pandas==2.2.2